refactor: move stats utils back to worker but outside of job runners
