+++ services/worker/src/worker/job_runners/split/duckdb_index.py
@@ -41,14 +41,14 @@ from libcommon.utils import HF_HUB_HTTP_ERROR_RETRY_SLEEPS, download_file_from_h
 
 from worker.config import AppConfig, DuckDbIndexConfig
 from worker.dtos import CompleteJobResult, SplitDuckdbIndex
-from worker.job_runners.split.descriptive_statistics import (
+from worker.job_runners.split.split_job_runner import SplitJobRunnerWithCache
+from worker.statistics_utils import (
     STRING_DTYPES,
     AudioColumn,
     ImageColumn,
     ListColumn,
     StringColumn,
 )
-from worker.job_runners.split.split_job_runner import SplitJobRunnerWithCache
 from worker.utils import (
     LOCK_GIT_BRANCH_RETRY_SLEEPS,
     create_branch,
